renamed agendaGroup to agendaGroups, also now able to sort items within their agenda containers
